Prova/App/requirements.txt
psycopg2-binary
pyyaml
Flask
prometheus-flask-exporter
flasgger
pandas
openpyxl
pandas
openpyxl